requirements.txt
asgiref==3.3.4
cpf-generator==2.1.4
deprecation==2.1.0
Django==3.2.3
djangorestframework==3.12.4
packaging==20.9
pyparsing==2.4.7
pytz==2021.1
sqlparse==0.4.1
